[Caching] Extract global-temp stable-key ordering into its own header
Move the content-keyed ordering subsystem (hstr/mix/local_key/stable_key/
store_value/sort_key plus its memo state) out of the hot, central
transforms/offload.cpp into a new self-contained GlobalTmpOrdering helper in
transforms/offload_stable_key.h. IdentifyValuesUsedInOtherOffloads keeps only
the traversal that collects cross-offload values and the per-alloca store map,
and assign_offsets() now delegates ordering to the helper before allocating
slots. No behavior change; addresses the feature-factorization CI check that
flagged piling this subsystem onto offload.cpp.
